requirements.txt
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
cython==3.0.12
decorator==5.2.1
fonttools==4.57.0
idna==3.10
importlib-resources==6.5.2
joblib==1.5.0
jsonpickle==4.0.5
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.43.0
madmom==0.16.1
matplotlib==3.9.4
mido==1.3.3
more-itertools==10.7.0
msgpack==1.1.0
music21==8.3.0
numba==0.60.0
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.7
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.1
scipy==1.13.1
setuptools==80.3.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
tslearn==0.6.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webcolors==24.11.1
zipp==3.21.0
python-rtmidi
rich
jsonschema
